feat(confirm-modal): marge files into a single file
